Fixed various compiler warnings.

+++ src/components/ogre/lod/ProgressiveMeshGenerator.cpp
@@ -660,7 +660,7 @@ void ProgressiveMeshGenerator::build(LodConfig& lodConfig)
 
 	// Lod distances needs to be sorted.
 	Ogre::Mesh::LodValueList values;
-	for (int i = 0; i < lodConfig.levels.size(); i++) {
+	for (size_t i = 0; i < lodConfig.levels.size(); i++) {
 		values.push_back(lodConfig.levels[i].distance);
 	}
 	mMesh->getLodStrategy()->assertSorted(values);
